Moved the chatbox as well to the main page and seperated the logic more so its easier to implement the sidebar
